[app] tabs added
