+++ lib/peer_tab.dart
@@ -1,0 +1,13 @@
+// peer_tab.dart
+import 'package:flutter/material.dart';
+
+class PeerTab extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      child: Center(
+        child: Text('Peer Tab'),
+      ),
+    );
+  }
+}
